This makes a number of changes to the motor positioning panel and control.
1 - A much reduced delay in motor response
2 - Separate buttons for zeroing either all axes (XYZA=0) or only XYZ (XYZ=0)
3 - Four separate memories (a,b,c and d) to store and retrieve positions (for XYZA)
4 - Fixes of bugs in the "rotate" and "pivot" configurations
5 - The pivotk variable in your config file is now gone and replaced by obj_length

In my case it reads:

sbconfig.obj_length = 98000;     % objective length from center of rotation to focal point [um]

The number should reflect the distance between the rotation axis of the objective to the focal point.  In my case it is 9.8cm

Make sure that positive rotation angles for the objective means it moves CW when seen from the front of the microscope -- otherwise things will not work in rotate and pivot modes.  (Lindsey -- wire back the objective motor direction the way it was originally.)

Position memory -- the left column is to store the current position and the right is to retrieve a stored position.  When retrieving the letter within the button will turn RED while the motors move and it will go back to black when the position has been retrieved.